Factor out eager val from eval_llama_lib
Would like to re-use EagerEvalWrapper and eval function for quantization calibration.

Differential Revision: [D57881028](https://our.internmc.facebook.com/intern/diff/D57881028/)

[ghstack-poisoned]
