+++ examples/models/llama2/evaluate/__init__.py
@@ -1,0 +1,12 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .eval import EagerEvalWrapper, evaluate_model
+
+__all__ = [
+    "evaluate_model",
+    "EagerEvalWrapper",
+]
